Trait sc_rpc_spec_v2::chain_head::api::ChainHeadApiClient

pub trait ChainHeadApiClient<Hash>: SubscriptionClientTwhere
    Hash: Send + Sync + 'static + Serialize + DeserializeOwned,{
    // Provided methods
    fn chain_head_unstable_body<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_header<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<String>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_genesis_hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_storage<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        hash: Hash,
        items: Vec<StorageQuery<String>>,
        child_trie: Option<String>
    ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_call<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        hash: Hash,
        function: String,
        call_parameters: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_unpin<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_continue<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        operation_id: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_stop_operation<'life0, 'async_trait>(
        &'life0 self,
        follow_subscription: String,
        operation_id: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn chain_head_unstable_follow<'life0, 'async_trait>(
        &'life0 self,
        with_runtime: bool
    ) -> Pin<Box<dyn Future<Output = Result<Subscription<FollowEvent<Hash>>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the ChainHeadApi RPC API.

Provided Methods

fn chain_head_unstable_body<'life0, 'async_trait>( &'life0 self, follow_subscription: String, hash: Hash ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Retrieves the body (list of transactions) of a pinned block.

This method should be seen as a complement to chainHead_unstable_follow, allowing the JSON-RPC client to retrieve more information about a block that has been reported.

Use archive_unstable_body if instead you want to retrieve the body of an arbitrary block.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_header<'life0, 'async_trait>( &'life0 self, follow_subscription: String, hash: Hash ) -> Pin<Box<dyn Future<Output = RpcResult<Option<String>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Retrieves the header of a pinned block.

This method should be seen as a complement to chainHead_unstable_follow, allowing the JSON-RPC client to retrieve more information about a block that has been reported.

Use archive_unstable_header if instead you want to retrieve the header of an arbitrary block.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_genesis_hash<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Get the chain’s genesis hash.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_storage<'life0, 'async_trait>( &'life0 self, follow_subscription: String, hash: Hash, items: Vec<StorageQuery<String>>, child_trie: Option<String> ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns storage entries at a specific block’s state.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_call<'life0, 'async_trait>( &'life0 self, follow_subscription: String, hash: Hash, function: String, call_parameters: String ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Call into the Runtime API at a specified block’s state.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_unpin<'life0, 'async_trait>( &'life0 self, follow_subscription: String, hash: Hash ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Unpin a block reported by the follow method.

Ongoing operations that require the provided block will continue normally.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_continue<'life0, 'async_trait>( &'life0 self, follow_subscription: String, operation_id: String ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Resumes a storage fetch started with chainHead_storage after it has generated an operationWaitingForContinue event.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_stop_operation<'life0, 'async_trait>( &'life0 self, follow_subscription: String, operation_id: String ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Stops an operation started with chainHead_unstable_body, chainHead_unstable_call, or chainHead_unstable_storage. If the operation was still in progress, this interrupts it. If the operation was already finished, this call has no effect.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_follow<'life0, 'async_trait>( &'life0 self, with_runtime: bool ) -> Pin<Box<dyn Future<Output = Result<Subscription<FollowEvent<Hash>>, Error>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Track the state of the head of the chain: the finalized, non-finalized, and best blocks.

Unstable

This method is unstable and subject to change in the future.