Trait sc_rpc_spec_v2::chain_head::api::ChainHeadApiServer

pub trait ChainHeadApiServer<Hash>: Sized + Send + Sync + 'static {
    // Required methods
    fn chain_head_unstable_body(
        &self,
        follow_subscription: String,
        hash: Hash
    ) -> RpcResult<MethodResponse>;
    fn chain_head_unstable_header(
        &self,
        follow_subscription: String,
        hash: Hash
    ) -> RpcResult<Option<String>>;
    fn chain_head_unstable_genesis_hash(&self) -> RpcResult<String>;
    fn chain_head_unstable_storage(
        &self,
        follow_subscription: String,
        hash: Hash,
        items: Vec<StorageQuery<String>>,
        child_trie: Option<String>
    ) -> RpcResult<MethodResponse>;
    fn chain_head_unstable_call(
        &self,
        follow_subscription: String,
        hash: Hash,
        function: String,
        call_parameters: String
    ) -> RpcResult<MethodResponse>;
    fn chain_head_unstable_unpin(
        &self,
        follow_subscription: String,
        hash: Hash
    ) -> RpcResult<()>;
    fn chain_head_unstable_continue(
        &self,
        follow_subscription: String,
        operation_id: String
    ) -> RpcResult<()>;
    fn chain_head_unstable_stop_operation(
        &self,
        follow_subscription: String,
        operation_id: String
    ) -> RpcResult<()>;
    fn chain_head_unstable_follow(
        &self,
        subscription_sink: SubscriptionSink,
        with_runtime: bool
    ) -> SubscriptionResult;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where Hash: Send + Sync + 'static + DeserializeOwned + Serialize { ... }
}

Server trait implementation for the ChainHeadApi RPC API.

Required Methods

fn chain_head_unstable_body( &self, follow_subscription: String, hash: Hash ) -> RpcResult<MethodResponse>

Retrieves the body (list of transactions) of a pinned block.

This method should be seen as a complement to chainHead_unstable_follow, allowing the JSON-RPC client to retrieve more information about a block that has been reported.

Use archive_unstable_body if instead you want to retrieve the body of an arbitrary block.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_header( &self, follow_subscription: String, hash: Hash ) -> RpcResult<Option<String>>

Retrieves the header of a pinned block.

This method should be seen as a complement to chainHead_unstable_follow, allowing the JSON-RPC client to retrieve more information about a block that has been reported.

Use archive_unstable_header if instead you want to retrieve the header of an arbitrary block.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_genesis_hash(&self) -> RpcResult<String>

Get the chain’s genesis hash.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_storage( &self, follow_subscription: String, hash: Hash, items: Vec<StorageQuery<String>>, child_trie: Option<String> ) -> RpcResult<MethodResponse>

Returns storage entries at a specific block’s state.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_call( &self, follow_subscription: String, hash: Hash, function: String, call_parameters: String ) -> RpcResult<MethodResponse>

Call into the Runtime API at a specified block’s state.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_unpin( &self, follow_subscription: String, hash: Hash ) -> RpcResult<()>

Unpin a block reported by the follow method.

Ongoing operations that require the provided block will continue normally.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_continue( &self, follow_subscription: String, operation_id: String ) -> RpcResult<()>

Resumes a storage fetch started with chainHead_storage after it has generated an operationWaitingForContinue event.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_stop_operation( &self, follow_subscription: String, operation_id: String ) -> RpcResult<()>

Stops an operation started with chainHead_unstable_body, chainHead_unstable_call, or chainHead_unstable_storage. If the operation was still in progress, this interrupts it. If the operation was already finished, this call has no effect.

Unstable

This method is unstable and subject to change in the future.

fn chain_head_unstable_follow( &self, subscription_sink: SubscriptionSink, with_runtime: bool ) -> SubscriptionResult

Track the state of the head of the chain: the finalized, non-finalized, and best blocks.

Unstable

This method is unstable and subject to change in the future.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where Hash: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.