sc_rpc_spec_v2
Module chain_spec

Substrate chain specification API.

The chain spec (short for chain specification) allows inspecting the content of the specification of the chain that a JSON-RPC server is targeting.

The values returned by the API are guaranteed to never change during the lifetime of the JSON-RPC server.

Note

Methods are prefixed by chainSpec.

Re-exports

pub use api::ChainSpecApiServer;

pub use chain_spec::ChainSpec;

Modules

api

API trait of the chain spec.

chain_spec

API implementation for the specification of a chain.