Trait sc_rpc_spec_v2::transaction::api::TransactionApiServer

pub trait TransactionApiServer<Hash: Clone>: Sized + Send + Sync + 'static {
    // Required method
    fn submit_and_watch(
        &self,
        subscription_sink: SubscriptionSink,
        bytes: Bytes
    ) -> SubscriptionResult;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where Hash: Send + Sync + 'static + Serialize { ... }
}

Server trait implementation for the TransactionApi RPC API.

Required Methods

fn submit_and_watch( &self, subscription_sink: SubscriptionSink, bytes: Bytes ) -> SubscriptionResult

Submit an extrinsic to watch.

See TransactionEvent for details on transaction life cycle.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where Hash: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.