Trait sc_rpc_spec_v2::transaction::api::TransactionBroadcastApiServer

pub trait TransactionBroadcastApiServer: Sized + Send + Sync + 'static {
    // Required methods
    fn broadcast(&self, bytes: Bytes) -> RpcResult<Option<String>>;
    fn stop_broadcast(&self, operation_id: String) -> Result<(), ErrorBroadcast>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the TransactionBroadcastApi RPC API.

Required Methods

fn broadcast(&self, bytes: Bytes) -> RpcResult<Option<String>>

Broadcast an extrinsic to the chain.

Unstable

This method is unstable and subject to change in the future.

fn stop_broadcast(&self, operation_id: String) -> Result<(), ErrorBroadcast>

Broadcast an extrinsic to the chain.

Unstable

This method is unstable and subject to change in the future.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Object Safety

This trait is not object safe.