Trait sc_rpc_spec_v2::archive::api::ArchiveApiClient

pub trait ArchiveApiClient<Hash>: ClientT
where
    Hash: Send + Sync + 'static + Serialize,
{
    // Provided methods
    fn archive_unstable_body<'life0, 'async_trait>(
        &'life0 self,
        hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<String>>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn archive_unstable_genesis_hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn archive_unstable_header<'life0, 'async_trait>(
        &'life0 self,
        hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<String>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn archive_unstable_finalized_height<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn archive_unstable_hash_by_height<'life0, 'async_trait>(
        &'life0 self,
        height: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<String>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn archive_unstable_call<'life0, 'async_trait>(
        &'life0 self,
        hash: Hash,
        function: String,
        call_parameters: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResult>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the ArchiveApi RPC API.

Provided Methods

fn archive_unstable_body<'life0, 'async_trait>( &'life0 self, hash: Hash ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<String>>>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Retrieves the body (list of transactions) of a given block hash.

Returns an array of strings containing the hexadecimal-encoded SCALE-codec-encoded transactions in that block. If no block with that hash is found, null.

Unstable

This method is unstable and subject to change in the future.

fn archive_unstable_genesis_hash<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<String>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Get the chain’s genesis hash.

Returns a string containing the hexadecimal-encoded hash of the genesis block of the chain.

Unstable

This method is unstable and subject to change in the future.

fn archive_unstable_header<'life0, 'async_trait>( &'life0 self, hash: Hash ) -> Pin<Box<dyn Future<Output = RpcResult<Option<String>>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Get the block’s header.

Returns a string containing the hexadecimal-encoded SCALE-codec encoding header of the block.

Unstable

This method is unstable and subject to change in the future.

fn archive_unstable_finalized_height<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Get the height of the current finalized block.

Returns an integer height of the current finalized block of the chain.

Unstable

This method is unstable and subject to change in the future.

fn archive_unstable_hash_by_height<'life0, 'async_trait>( &'life0 self, height: u64 ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<String>>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Get the hashes of blocks from the given height.

Returns an array (possibly empty) of strings containing an hexadecimal-encoded hash of a block header.

Unstable

This method is unstable and subject to change in the future.

fn archive_unstable_call<'life0, 'async_trait>( &'life0 self, hash: Hash, function: String, call_parameters: String ) -> Pin<Box<dyn Future<Output = RpcResult<MethodResult>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Call into the Runtime API at a specified block’s state.

Unstable

This method is unstable and subject to change in the future.

Object Safety

This trait is not object safe.