Module sc_rpc_spec_v2::chain_head::event

The chain head’s event returned as json compatible object.

Structs

BestBlockChanged

Indicate the block hash of the new best block.

ChainHeadRuntimeVersion

Simplified type clone of sp_version::RuntimeVersion. Used instead of sp_version::RuntimeVersion to conform to RPC spec V2.

ErrorEvent

The operation could not be processed due to an error.

Finalized

Indicate the finalized and pruned block hashes.

Initialized

Contain information about the latest finalized block.

MethodResponseStarted

The started result of a method.

NewBlock

Indicate a new non-finalized block.

OperationBodyDone

The response of the chainHead_body method.

OperationCallDone

The response of the chainHead_call method.

OperationError

Indicate a problem during the operation.

OperationId

Indicate the operation id of the event.

OperationStorageItems

The response of the chainHead_call method.

RuntimeVersionEvent

The runtime specification of the current block.

StorageQuery

The storage item received as paramter.

StorageResult

The storage result.

Enums

FollowEvent

The event generated by the follow method.

MethodResponse

The method respose of chainHead_body, chainHead_call and chainHead_storage.

RuntimeEvent

The runtime event generated if the follow subscription has set the with_runtime flag.

StorageQueryType

The type of the storage query.

StorageResultType

The type of the storage query.