Module sc_rpc_spec_v2::chain_head::event

The chain head’s event returned as json compatible object.

Structs

BestBlockChanged

Indicate the block hash of the new best block.

ChainHeadResult

The result of a chain head method.

ErrorEvent

The operation could not be processed due to an error.

Finalized

Indicate the finalized and pruned block hashes.

Initialized

Contain information about the latest finalized block.

NetworkConfig

The network config parameter is used when a function needs to request the information from its peers.

NewBlock

Indicate a new non-finalized block.

RuntimeVersionEvent

The runtime specification of the current block.

Enums

ChainHeadEvent

The event generated by the body / call / storage methods.

FollowEvent

The event generated by the follow method.

RuntimeEvent

The runtime event generated if the follow subscription has set the runtime_updates flag.