sc_rpc_spec_v2

Module transaction

Substrate transaction API.

The transaction methods allow submitting a transaction and subscribing to its status updates generated by the chain.

Note

Methods are prefixed by transaction.

Re-exports

pub use api::TransactionApiServer;

pub use api::TransactionBroadcastApiServer;

pub use event::TransactionBlock;

pub use event::TransactionDropped;

pub use event::TransactionError;

pub use event::TransactionEvent;

pub use transaction::Transaction;

pub use transaction_broadcast::TransactionBroadcast;

Modules

api

API trait for transactions.

error

Transaction RPC errors.

event

The transaction’s event returned as json compatible object.

transaction

API implementation for submitting transactions.

transaction_broadcast

API implementation for broadcasting transactions.

Structs

TransactionMetrics

RPC layer metrics for transaction pool.