Module sc_rpc_spec_v2::transaction

Substrate transaction API.

The transaction methods allow submitting a transaction and subscribing to its status updates generated by the chain.

Note

Methods are prefixed by transaction.

Re-exports

pub use api::TransactionApiServer;

pub use event::TransactionBlock;

pub use event::TransactionBroadcasted;

pub use event::TransactionDropped;

pub use event::TransactionError;

pub use event::TransactionEvent;

pub use transaction::Transaction;

Modules

api

API trait for transactions.

error

Transaction RPC errors.

event

The transaction’s event returned as json compatible object.

transaction

API implementation for submitting transactions.